use method call to load usernames and cache results
After this change username updates will be cached and won't update
reactively. This is no big deal because usernames change rarely.

This homegrown caching solution leaves me a bit uneasy. Couldn't be
bothered to look for a good memoizer though.

The inital reason to change the loading meachnism is that SubsCache uses a
timeout when subscribing which is not allowed from inside stubs. Apparently
rebuilding the template happens as part of the simulation(?) so that
it would try to start a subscription from inside the stub. Wild guessing
on my part. Anyway, this fixes #548