Add support for web analytics using Matomo, fixes #855
Does not do anything unless `public.matomo` setting is populated with the
following properties:

```
"matomo": {
    "url": "http://analytics.example.com/",
    "site": 1
}
```

Where `url` is the base url to the matomo instance (terminating slash is
important) and `site` is the numeric id of the site configured on that
matomo instance.

Iron-router integration was implemented in accordance with
recommendations for single page applications published in the
[matomo blog][1].

Regrettably there is currently a [matomo bug][2] preventing reliable
user-tracking in single-page-applications, hence this is not supported
at this time.

[1]: https://matomo.org/blog/2017/02/how-to-track-single-page-websites-using-piwik-analytics/
[2]: https://github.com/matomo-org/matomo/issues/7556